% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{add_line_labels}
\alias{add_line_labels}
\title{Add labels to line graphs}
\usage{
add_line_labels(y_var, percent = FALSE)
}
\description{
Add labels to line graphs
}
\examples{
p <- ggplot(mtcars, aes(gear, hp)) + geom_bar(stat="identity") + add_bar_labels("hp", percent = TRUE)
add_line_labelss()
}
