% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{add_stacked_bar_labels}
\alias{add_stacked_bar_labels}
\title{Add labels to stacked bargraphs}
\usage{
add_stacked_bar_labels(y_var)
}
\description{
Add labels to stacked bargraphs
}
\examples{
p <- ggplot(mtcars, aes(gear, hp)) + geom_bar(stat="identity") + add_stacked_bar_labels(percent = TRUE)
add_stacked_bar_labels()
}
