% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{colour_wiki_continuous}
\alias{colour_wiki_continuous}
\title{Setting scale functions}
\usage{
colour_wiki_continuous(name, n = NULL)
}
\arguments{
\item{name}{A character string indicating the name of the palette to use from the \code{wiki_colors} list.}
}
\value{
A ggplot2 scale object for continuous color palettes.
}
\description{
This function creates a ggplot2 scale for continuous color palettes.
}
\examples{
colour_wiki_continuous()
}
