% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{colour_wiki_discrete}
\alias{colour_wiki_discrete}
\title{Setting scale functions}
\usage{
colour_wiki_discrete(name, n = NULL)
}
\arguments{
\item{name}{A character string indicating the name of the palette to use from the \code{wiki_colors} list.}
}
\value{
A ggplot2 scale object for discrete color palettes.
}
\description{
This function creates a ggplot2 scale for discrete color palettes.
}
\examples{
color_wiki_discrete()
}
