% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{.onDetach}
\alias{.onDetach}
\title{Resets the settings when the package is detached, experimental.}
\usage{
.onDetach(libname)
}
\description{
Resets the settings when the package is detached, experimental.
}
\examples{
.onDetach("mypkg")
.onDetach()
}
