% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{.onLoad}
\alias{.onLoad}
\title{Sets up the desired font for showtext package}
\usage{
.onLoad(libname, pkgname)
}
\description{
Sets up the desired font for showtext package
}
\examples{
.onLoad("mypkg")
.onLoad()
}
