% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{fill_wiki_continuous}
\alias{fill_wiki_continuous}
\title{Setting scale functions}
\usage{
fill_wiki_continuous(name)
}
\arguments{
\item{name}{A character string indicating the name of the palette to use from the \code{wiki_colors} list.}
}
\value{
A ggplot2 scale object for continuous fill palettes.
}
\description{
This function creates a ggplot2 scale for continuous fill palettes.
}
\examples{
fill_wiki_continuous()
}
