% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{fill_wiki_discrete}
\alias{fill_wiki_discrete}
\title{Setting scale functions}
\usage{
fill_wiki_discrete(name, n = NULL)
}
\arguments{
\item{name}{A character string indicating the name of the palette to use from the \code{wiki_colors} list.}
}
\value{
A ggplot2 scale object for discrete fill palettes.
}
\description{
This function creates a ggplot2 scale for discrete fill palettes.
}
\examples{
fill_wiki_discrete()
}
