% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{ggplot_add.theme_wiki}
\alias{ggplot_add.theme_wiki}
\title{Removes Legend if plot meets condition}
\usage{
\method{ggplot_add}{theme_wiki}(object, plot, object_name)
}
\description{
This creates wiki theme for ggplot2
}
