% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwiki-package.R
\docType{package}
\name{ggwiki}
\alias{ggwiki}
\alias{ggwiki-package}
\title{ggwiki: An R Package for Customizing ggplot2 Plots with a Wiki Theme}
\description{
The ggwiki package provides custom themes, scales, and labeling functions
for ggplot2 to create plots with a wiki-style appearance. It includes functions
for customizing the appearance of plots, adding labels, and wrapping axis labels.
The main functions are \code{theme_wiki()}, \code{add_bar_labels()}, \code{x_wrap_discrete()},
\code{y_wrap_cont()}, \code{y_wrap_discrete()}, and \code{x_wrap_cont()}.
}
