% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{scale_colour_wiki_c}
\alias{scale_colour_wiki_c}
\title{Setting scale functions}
\usage{
scale_colour_wiki_c(name)
}
\arguments{
\item{name}{A character string indicating the name of the palette to use from the \code{wiki_colors} list.}
}
\value{
A ggplot2 scale object for continuous color palettes.
}
\description{
This function creates a ggplot2 scale for continuous color palettes.
}
\examples{
scale_colour_wiki_c("main")
}
