% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{theme_wiki}
\alias{theme_wiki}
\title{Creates a custom "wiki" theme for ggplot2}
\usage{
theme_wiki(..., simple = FALSE)
}
\description{
Creates a custom "wiki" theme for ggplot2
}
\examples{
p <- ggplot(mtcars, aes(mpg, hp)) + geom_point() + theme_wiki()
theme_wiki()
}
