% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{wiki_palettes}
\alias{wiki_palettes}
\title{Setting scale functions}
\usage{
wiki_palettes(
  name,
  n,
  all_palettes = wiki_colors,
  type = c("discrete", "continuous")
)
}
\description{
This creates scales to be used with ggplot2
}
\examples{
wiki_palettes()
}
