% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{y_wrap_cont}
\alias{y_wrap_cont}
\title{Custom scale_y_continuous function that wraps long labels}
\usage{
y_wrap_cont(width = 12, percent = FALSE, ...)
}
\description{
Custom scale_y_continuous function that wraps long labels
}
\examples{
p <- ggplot(mtcars, aes(mpg, hp)) + geom_point() + y_wrap_cont(width = 10)
y_wrap_cont()
}
