% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wiki.R
\name{y_wrap_discrete}
\alias{y_wrap_discrete}
\title{Custom scale_y_discrete function that wraps long labels}
\usage{
y_wrap_discrete(width = 12, ...)
}
\description{
Custom scale_y_discrete function that wraps long labels
}
\examples{
p <- ggplot(mtcars, aes(mpg, gear)) + geom_point() + y_wrap_discrete(width = 10)
y_wrap_discrete()
}
