% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_palettes.R
\name{scale_fill_wiki_d}
\alias{scale_fill_wiki_d}
\title{Setting scale functions}
\usage{
scale_fill_wiki_d(name)
}
\description{
This creates scales to be used with ggplot2
}
\examples{
scale_fill_wiki_d()
}
